refactor(silo): move test files next to source files and standardize naming
